util.c: do not set  SO_RCVBUF
The original purpose of setting SO_RCVBUF as stated in the comment was to avoid packet loss, but on high bitrates (e.g. several hundreds of megabits) the newly set buffer size itself becomes the chokepoint, causing packet loss and defeating the original purpose.

Since there's no way to control the buffer size externally or via command-line, avoid setting it altogether and let the user control it via procfs.